--- pkg/plugin/docker.go
@@ -1,336 +1,0 @@
-// Package plugin provides an easy way to create the required CLI for a plugin.
-// It abstracts away all the command line and file parsing so users just need to implement the actual logic.
-package plugin
-
-import (
-	"context"
-
-	"fmt"
-	"strings"
-	"time"
-
-	"github.com/Blockdaemon/bpm-sdk/pkg/docker"
-	"github.com/Blockdaemon/bpm-sdk/pkg/node"
-	"github.com/Blockdaemon/bpm-sdk/pkg/template"
-
-	homedir "github.com/mitchellh/go-homedir"
-)
-
-// DockerPlugin is a varation of Plugin that comes with default methods for docker
-type DockerPlugin struct {
-	// All configurations files and their respective templates
-	configFilesAndTemplates map[string]string
-
-	// All containers that should be managed by this plugin
-	containers []docker.Container
-
-	// Plugin meta information
-	meta MetaInfo
-
-	name string
-}
-
-const (
-	filebeatContainerImage = "docker.elastic.co/beats/filebeat:7.4.1"
-	filebeatContainerName  = "filebeat"
-	filebeatConfigFile     = "filebeat.yml"
-)
-
-func NewDockerPlugin(name string, containers []docker.Container, configFilesAndTemplates map[string]string, meta MetaInfo) DockerPlugin {
-	// Add filebeat to the passed in containers
-	filebeatContainer := docker.Container{
-		Name:  filebeatContainerName,
-		Image: filebeatContainerImage,
-		Cmd:   []string{"-e", "-strict.perms=false"},
-		// using the first containers network is a decent default, if we ever do mult-network deployments we may need to rethink this
-		NetworkID: containers[0].NetworkID,
-		Mounts: []docker.Mount{
-			{
-				Type: "bind",
-				From: filebeatConfigFile,
-				To:   "/usr/share/filebeat/filebeat.yml",
-			},
-			{
-				Type: "bind",
-				From: "/var/lib/docker/containers",
-				To:   "/var/lib/docker/containers",
-			},
-		},
-		User: "root",
-	}
-
-	containers = append(containers, filebeatContainer)
-	configFilesAndTemplates[filebeatConfigFile] = filebeatConfigTpl
-
-	return DockerPlugin{
-		configFilesAndTemplates: configFilesAndTemplates,
-		containers:              containers,
-		name:                    name,
-		meta:                    meta,
-	}
-}
-
-func (d DockerPlugin) Name() string {
-	return d.name
-}
-
-func (d DockerPlugin) Meta() MetaInfo {
-	return d.meta
-}
-
-func (d DockerPlugin) ValidateParameters(currentNode node.Node) error {
-	for _, parameter := range d.Meta().Parameters {
-		ok := false
-
-		if parameter.Type == ParameterTypeBool {
-			_, ok = currentNode.BoolParameters[parameter.Name]
-		}
-
-		if parameter.Type == ParameterTypeString {
-			_, ok = currentNode.StrParameters[parameter.Name]
-
-		}
-
-		if !ok {
-			return fmt.Errorf(`%q missing`, parameter.Name)
-		}
-	}
-
-	return nil
-}
-
-// CreateSecrets does nothing except printing that it does nothing
-func (d DockerPlugin) CreateSecrets(currentNode node.Node) error {
-	if err := d.ValidateParameters(currentNode); err != nil {
-		return err
-	}
-	fmt.Println("Nothing to do here, skipping create-secrets")
-	return nil
-}
-
-// The default upgrade strategy removes all containers. If they where running they get started again which will pull new container images.
-//
-// This works as long as only the container version changes. If the the upgrade needs changes to the configs or migrations tasks it is
-// recommended to overwrite this function
-func (d DockerPlugin) Upgrade(currentNode node.Node) error {
-	status, err := d.Status(currentNode)
-	if err != nil {
-		return err
-	}
-
-	// Remove old containers so that the next start will pull the new ones
-	if err := d.RemoveRuntime(currentNode); err != nil {
-		return err
-	}
-
-	// If node was running, start it again
-	if status == "running" {
-		if err := d.Start(currentNode); err != nil {
-			return err
-		}
-	}
-
-	return nil
-}
-
-// Test does nothing except printing that it does nothing
-func (d DockerPlugin) Test(currentNode node.Node) (bool, error) {
-	fmt.Println("Nothing to do here, skipping test")
-	return true, nil
-}
-
-// CreateConfigs creates configuration files for the blockchain client
-func (d DockerPlugin) CreateConfigs(currentNode node.Node) error {
-	if err := d.ValidateParameters(currentNode); err != nil {
-		return err
-	}
-	return template.ConfigFilesRendered(d.configFilesAndTemplates, template.TemplateData{
-		Node: currentNode,
-		PluginData: map[string]interface{}{
-			"Containers": d.containers,
-		},
-	})
-}
-
-// Start starts monitoring agents and delegates to another function to start blockchain containers
-func (d DockerPlugin) Start(currentNode node.Node) error {
-	client, err := docker.NewBasicManager(currentNode.NamePrefix(), currentNode.ConfigsDirectory())
-	if err != nil {
-		return err
-	}
-
-	ctx, cancel := context.WithTimeout(context.Background(), 2*time.Minute)
-	defer cancel()
-
-	// First, create the docker network(s) if they don't exist yet
-	for _, container := range d.containers {
-		if err := client.NetworkExists(ctx, container.NetworkID); err != nil {
-			return err
-		}
-	}
-
-	//////////////////////////////////////////////////////////////////////////////
-	// TODO: This is just temporarily until we have a proper authentication system
-	//////////////////////////////////////////////////////////////////////////////
-	if currentNode.Collection.Host != "" {
-		currentNode.Collection.Key, err = homedir.Expand(currentNode.Collection.Key)
-		if err != nil {
-			return err
-		}
-		currentNode.Collection.Cert, err = homedir.Expand(currentNode.Collection.Cert)
-		if err != nil {
-			return err
-		}
-		currentNode.Collection.CA, err = homedir.Expand(currentNode.Collection.CA)
-		if err != nil {
-			return err
-		}
-
-		for ix, container := range d.containers {
-			if strings.HasSuffix(container.Name, "beat") { // yeah, I know, super hacky but it's just temporarily
-				fmt.Printf("Add ssl certs to container: %s\n", container.Name)
-				sslMounts := []docker.Mount{
-					{
-						Type: "bind",
-						From: currentNode.Collection.CA,
-						To:   "/etc/ssl/beats/ca.crt",
-					},
-					{
-						Type: "bind",
-						From: currentNode.Collection.Cert,
-						To:   "/etc/ssl/beats/beat.crt",
-					},
-					{
-						Type: "bind",
-						From: currentNode.Collection.Key,
-						To:   "/etc/ssl/beats/beat.key",
-					},
-				}
-				d.containers[ix].Mounts = append(container.Mounts, sslMounts...)
-			}
-		}
-	}
-	//////////////////////////////////////////////////////////////////////////////
-	// TODO end
-	//////////////////////////////////////////////////////////////////////////////
-
-	// Next, start the containers
-	for _, container := range d.containers {
-		if err := client.ContainerRuns(ctx, container); err != nil {
-			return err
-		}
-	}
-
-	return nil
-}
-
-// DockerStatus returns the status of the running blockchain client and monitoring containers
-func (d DockerPlugin) Status(currentNode node.Node) (string, error) {
-	client, err := docker.NewBasicManager(currentNode.NamePrefix(), currentNode.ConfigsDirectory())
-	if err != nil {
-		return "", err
-	}
-
-	ctx, cancel := context.WithTimeout(context.Background(), 2*time.Minute)
-	defer cancel()
-
-	containersRunning := 0
-
-	for _, container := range d.containers {
-		running, err := client.IsContainerRunning(ctx, container.Name)
-		if err != nil {
-			return "", err
-		}
-		if running {
-			containersRunning += 1
-		}
-	}
-
-	if containersRunning == 0 {
-		return "stopped", nil
-	} else if len(d.containers) == containersRunning {
-		return "running", nil
-	}
-
-	return "incomplete", nil
-}
-
-// DockerStop removes all containers
-func (d DockerPlugin) Stop(currentNode node.Node) error {
-	client, err := docker.NewBasicManager(currentNode.NamePrefix(), currentNode.ConfigsDirectory())
-	if err != nil {
-		return err
-	}
-
-	ctx, cancel := context.WithTimeout(context.Background(), 2*time.Minute)
-	defer cancel()
-
-	for _, container := range d.containers {
-		if err = client.ContainerStopped(ctx, container.Name, container.NetworkID); err != nil {
-			return err
-		}
-	}
-
-	return nil
-}
-
-// Removes any data (typically the blockchain itself) related to the node
-func (d DockerPlugin) RemoveData(currentNode node.Node) error {
-	client, err := docker.NewBasicManager(currentNode.NamePrefix(), currentNode.ConfigsDirectory())
-	if err != nil {
-		return err
-	}
-
-	ctx, cancel := context.WithTimeout(context.Background(), 2*time.Minute)
-	defer cancel()
-	// Remove volumes
-	for _, container := range d.containers {
-		for _, mount := range container.Mounts {
-			if mount.Type == "volume" {
-				if err = client.VolumeAbsent(ctx, mount.From); err != nil {
-					return err
-				}
-			}
-		}
-	}
-
-	return nil
-}
-
-// Removes configuration files related to the node
-func (d DockerPlugin) RemoveConfig(currentNode node.Node) error {
-	// Remove all configuration files
-	for file := range d.configFilesAndTemplates {
-		if err := template.ConfigFileAbsent(file, currentNode); err != nil {
-			return err
-		}
-	}
-
-	return nil
-}
-
-// Removes the docker network and containers
-func (d DockerPlugin) RemoveRuntime(currentNode node.Node) error {
-	client, err := docker.NewBasicManager(currentNode.NamePrefix(), currentNode.ConfigsDirectory())
-	if err != nil {
-		return err
-	}
-
-	ctx, cancel := context.WithTimeout(context.Background(), 4*time.Minute)
-	defer cancel()
-
-	for _, container := range d.containers {
-		if err = client.ContainerAbsent(ctx, container.Name, container.NetworkID); err != nil {
-			return err
-		}
-	}
-
-	// Remove network(s)
-	for _, container := range d.containers {
-		if err := client.NetworkAbsent(ctx, container.NetworkID); err != nil {
-			return err
-		}
-	}
-
-	return nil
-}
